input binding slowly starts to work, but it's very hacked. i will need to rewrite it at some point

+++ animFlex/source/AFStaticMeshComponent.cpp
@@ -23,7 +23,7 @@ void AFStaticMeshComponent::Draw() const
 
 int AFStaticMeshComponent::GetTrianglesNum()
 {
-	return m_mesh.vertices.size() / 3;
+	return static_cast<int>(m_mesh.vertices.size() / 3);
 }
 
 bool AFStaticMeshComponent::Load()
